PSR-266 Renamed controller
