+++ conf/app.routes
@@ -280,14 +280,14 @@ GET         /:srn/submit-file-upload-error-missing-information                  
 GET         /:srn/change-file-upload-error-missing-information                          controllers.nonsipp.memberdetails.upload.FileUploadErrorMissingInformationController.onPageLoad(srn: Srn, mode: Mode = CheckMode)
 GET         /:srn/change-submit-file-upload-error-missing-information                   controllers.nonsipp.memberdetails.upload.FileUploadErrorMissingInformationController.onSubmit(srn: Srn, mode: Mode = CheckMode)
 
-GET         /:srn/type-of-:subject                                                      controllers.nonsipp.common.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999 = 1, mode: Mode = NormalMode, subject: IdentitySubject)
-GET         /:srn/type-of-:subject/:index                                               controllers.nonsipp.common.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999, mode: Mode = NormalMode, subject: IdentitySubject)
-POST        /:srn/type-of-:subject                                                     controllers.nonsipp.common.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999 = 1, mode: Mode = NormalMode, subject: IdentitySubject)
-POST        /:srn/type-of-:subject/:index                                              controllers.nonsipp.common.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999, mode: Mode = NormalMode, subject: IdentitySubject)
-GET         /:srn/change-type-of-:subject                                               controllers.nonsipp.common.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999 = 1, mode: Mode = CheckMode, subject: IdentitySubject)
-GET         /:srn/change-type-of-:subject/:index                                        controllers.nonsipp.common.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999, mode: Mode = CheckMode, subject: IdentitySubject)
-POST        /:srn/change-type-of-:subject                                              controllers.nonsipp.common.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999 = 1, mode: Mode = CheckMode, subject: IdentitySubject)
-POST        /:srn/change-type-of-:subject/:index                                       controllers.nonsipp.common.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999, mode: Mode = CheckMode, subject: IdentitySubject)
+GET         /:srn/type-of-:subject                                                      controllers.nonsipp.common.IdentityTypeController.onPageLoad(srn: Srn, index: Max9999999 = 1, mode: Mode = NormalMode, subject: IdentitySubject)
+GET         /:srn/type-of-:subject/:index                                               controllers.nonsipp.common.IdentityTypeController.onPageLoad(srn: Srn, index: Max9999999, mode: Mode = NormalMode, subject: IdentitySubject)
+POST        /:srn/type-of-:subject                                                     controllers.nonsipp.common.IdentityTypeController.onSubmit(srn: Srn, index: Max9999999 = 1, mode: Mode = NormalMode, subject: IdentitySubject)
+POST        /:srn/type-of-:subject/:index                                              controllers.nonsipp.common.IdentityTypeController.onSubmit(srn: Srn, index: Max9999999, mode: Mode = NormalMode, subject: IdentitySubject)
+GET         /:srn/change-type-of-:subject                                               controllers.nonsipp.common.IdentityTypeController.onPageLoad(srn: Srn, index: Max9999999 = 1, mode: Mode = CheckMode, subject: IdentitySubject)
+GET         /:srn/change-type-of-:subject/:index                                        controllers.nonsipp.common.IdentityTypeController.onPageLoad(srn: Srn, index: Max9999999, mode: Mode = CheckMode, subject: IdentitySubject)
+POST        /:srn/change-type-of-:subject                                              controllers.nonsipp.common.IdentityTypeController.onSubmit(srn: Srn, index: Max9999999 = 1, mode: Mode = CheckMode, subject: IdentitySubject)
+POST        /:srn/change-type-of-:subject/:index                                       controllers.nonsipp.common.IdentityTypeController.onSubmit(srn: Srn, index: Max9999999, mode: Mode = CheckMode, subject: IdentitySubject)
 
 GET         /:srn/file-upload-error-some                                                controllers.nonsipp.memberdetails.upload.FileUploadErrorSummaryController.onPageLoad(srn: Srn, mode: Mode = NormalMode)
 GET         /:srn/submit-file-upload-error-some                                         controllers.nonsipp.memberdetails.upload.FileUploadErrorSummaryController.onSubmit(srn: Srn, mode: Mode = NormalMode)
